Add basic bookmark support
Whenever you open Oni from the menu or clicking on the icon, it can
cause the fuzzy finder to try to find through-out the install directory,
which made Oni unusable and have to crash it or kill it via task manager
or kill -9 on linux, etc.
This checks if you are in the install directory, if so then open
bookmarks instead of searching through files.

- Implement oni.bookmarks so user can set these.
- If you click loading indicator it does nothing.
- Comment some sections of the code I have concerns about.
- Add different icon type for bookmarks.
- If bookmark change directory of vim to the bookmark
  (which triggers DirChanged so also Oni.)

TODO: This really needs to wait until we have an "Open Folder" solution,
so that we know how to better handle bookmarks and how to combine that
with opening folders.