Fix for #549, async git ls-files
- Remove getUntrackedFiles, getTrackedFiles, and isGitDir due to them
not being used anywhere else and this can all be accomplished with one
call rather than three.

- Convert overriden method and git ls-files to async calls.

TODO: Still need to have a binary or something to async load files in
normal situations, it's a bit better now but it's still not ideal.

I wrote a python script which converted to binary and it takes command
line options such as ignore extensions, ignore directory names, and
ignore hidden directories. We can use it to either list files, or
directories async (tested locally on windows and linux (different
executables), that way it's all the same function for show() in
QuickOpen, which reduces code size, and import size.