Fix tabs not changedTick instead of modified
If one would modify a buffer then undo it, instead of changing from
isDirty to not dirty (tabline with the little circle) it would remain
dirty because ticks update even when undoing...

Therefore check for modified instead.

This would have caused furthur issues whenever we implement
requesting not close on dirty.

Issue(s) : #624