Speed enhancements and caching
- Ignore image extensions, need to add more and more binary.. maybe
convert to json? Idk strings are pretty easy.

- Fixed ignore for git files so we ignore binary and image as well as
global git ignore, need to add .oniignore, but maybe not? We have it
already as a setting do we really need another config file?

- Re-filter input if still loading directories (This is a big edge case)

- Cache files on dir changed and construction, that way if we don't move
dirs which most of the times people don't, it will just have the files
saved, ready for loading.

- Manual walk enhanced.

Noticed a bug in overriden commands, doesn't seem to work at all with
ag, need to investigate if it's just ag -l --nocolor or if it's others
as well, git works fine... weird.