Optimistic typing updates (#946)
One consequence of the architecture we have - Electron talking to Neovim via `msgpack-rpc` - is that input has a non-trivial round-trip latency, regardless of the configuration.

For an editor, ideally you'd want an avg responsiveness of <10ms from keypress to rendering for a desirable experience (ex, [Typing With Pleasure](https://pavelfatin.com/typing-with-pleasure)). Avg of 15ms is pretty good.

At least based on the chrome profiler, we aren't there yet - and we spend a significant amount of time idling until we get a response from Neovim, in which case a bunch of stuff happens - lots of redux actions in response to mouse-move, canvas re-rendering, etc.

This change implements 'optimistic updates' - similiar to the IntelliJ IDEA zero-latency (it also has parallels to [Client-side network prediction](https://developer.valvesoftware.com/wiki/Source_Multiplayer_Networking) models used in multiplayer games, just a really simple version!).

For insert mode, we can predict fairly reliable on the cursor position + character, so we should render that _immediately_.

Not only does this mean that we achieve better typing latency, but it also amortizes the workload in a more efficient way (since the cursor already moved, and lots of components get re-rendered in response to that, the canvas rendering can occur more quickly too).