Bugfix/ Fix staggered cursor positioner rendering (#2285)
This PR fixes what is arguably a small but potentially (at least for me *incredibly frustrating*) bug which is that when the hover renderer appears depending on where it is it seems to shuffle into place aka renders in one position then repositions v. quickly before settling into place 💢. Hopefully this PR helps resolve this frustration for anyone else who is as finicky as me re. such things.

Before:

![broken_hover](https://user-images.githubusercontent.com/22454918/40940777-c79b1810-6840-11e8-9754-61c228159e36.gif)

After:

![fixed_hover](https://user-images.githubusercontent.com/22454918/40940791-ce466552-6840-11e8-8c7f-7be573aa3fcc.gif)