Fix test conflict
* `epicMiddleware.replaceEpic` internally dispatches a
`@@redux-observable/EPIC_END` action, which gets saved to the
`mockStore`'s
actions list. Every test case's `afterEach` thus adds an additional
action to this actions list. This pollutes all subsequent stores created
via the `mockStore` factory, which is odd.
* This was accounted for in the YANK_AND_PASTE_EPICS test by assuming
"an init action is sent first", which is not the case.
* So move the setup that is only for the SET_ROOT_DIRECTORY test inside
it's own `describe` block, remove the unnecessary `replaceEpic` call,
and fix the test for expected number of actions.