Feature/ Indent lines improvements (#2446)
As mentioned in #2344 buffer layers suffer from laggy rendering on scroll, on some investigation it seems the `auditTime` we use in the `shouldMeasure$` observable means that the information is not being updated in real time so scrolling quickly through a buffer causes indentlines or color highlights to appear in a *janky* fashion i.e. they are out of place.

Not sure what it was used for but I think because the information is required by ui pieces which need v. up to date data re window details we should allow things to be updated as often possible. This change greatly improves the rendering.

There are a couple of things which still cause the ui to be a bit janky:
* one is the initial buffer entering although this is arguably low priority 🤷‍♂️ - seems quite a lot happens there the first time and rendering is noticeably slower.
* folding is completely unhandled (not investigated yet)
* double line wrapping
* searching in a buffer which can change the position of the window doesn't seem to update the window positioning (any ideas @bryphe?) - i.e. searching in buffer with `?` or `/` can move the viewport towards the match(es) in this case window details passed to context don't seem to update
* initial load of a buffer similar to one
### Other additions:

* I also tweaked the indent lines to use a `bannedFiletypes` buffer as opposed to a whitelist aka as discussed with @CrossR it is now on for all filetypes except those which are blacklisted. @CrossR examples of where you would want a black list I think would be most vim plugin buffers like `peekabo` or `startify` or `nerdtree` etc. in these cases the indentlines rarely render desirably and are arguably not required.

* I render the first indent line which is the same behaviour as in `vscode` however i've added an option. `experimental.indentLines.skipFirst` which allows a user to avoid this behaviour if they would prefer