Bugfix/fix welcome buffer layer (#2522)
This PR will fix #1413 and #1459 it uses the handle input method on the buffer layers to define a custom input handler for the welcome layer which is fairly simple i.e. `j` and `k` for navigation and `<enter>` to select, this PR also adds a method to the oni api called `getActiveSection` which returns a string representing the active section of the editor e.g. the particular sidebar element or the menu or commandline

Outstanding

- [x] Hide cursor when welcome is active
- [x] Fix welcome commands (some...)
- [x] Allow `enter` key passthrough
- [x] Add tests

~add sessions to welcome screen~ leave for another PR