Bugfix/Fix dynamic import for shell env (#2640)
The `shell env` package has been silently failing meaning that fixes that allow oni to read in a user's env var have regressed. This happened when we upgraded `webpack` because with the new version of webpack, dynamic imports now return objects with the shape `{ default: DefaultFnExport, aVar: var }`

Also upgraded the `shellenv` package as is since been upgraded.